Still on the way
Getting gobbledygook from the remote end when it receives data from the
replaying end. This is (probably) an issue with the sending end rather
than the receiver because I’ve got the receiver parsing the raw data
into  a string and echoing it out faithfully.

-> Look into what the relay code is doing - probably an issue with a
c_str() or something somewhere.